backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.1
mercadopago==2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
